Revert "* Method getRootDirectory() is used in CUnit class"
stpreproc needn't necessarily be along with stanse.jar in common
directory.

Add dist/ into the $PATH if you want (even internally), but NOTE, the
scripts will no longer be with stanse.jar in the same directory when
packaged into a distribution. stanse.jar will reside in
/usr/*/share/java/ whereas scripts in /usr/*/bin. There is no point
to care about where scripts are.

git-svn-id: https://arran.fi.muni.cz/svn/stanse/trunk@461 98efdb2a-0913-0410-9d53-f03ad6a1d54d