cparser: speedup and fixup initializers
Performance of initializers was poor and what is worse, initializers
conversion to AST was buggy.

On a 42-lined file full of initializers the parsing took 5 s, now it's
finished in 600 ms. There was a misunderstanding how Antlr handles star.
Fix that shit.

Now it doesn't crash on OOM when parsing files with initializers.

git-svn-id: https://arran.fi.muni.cz/svn/stanse/trunk@592 98efdb2a-0913-0410-9d53-f03ad6a1d54d