configuration: do filtering on the unit level
We got rid of error filtering, but we still should do some. Now we can
do it so that we always report the shortest trace ever :).

Also there is a possibility to count z-ranking. The code is
more or less there, but commented. It needs a fix in that it doesn't
have unique keys in the map. Finally the rank should be reported
somehow.

git-svn-id: https://arran.fi.muni.cz/svn/stanse/trunk@1147 98efdb2a-0913-0410-9d53-f03ad6a1d54d